Adjustments for CastIron 0.3.1

+++ src/game_assets/hex_grid_cell.rs
@@ -107,7 +107,7 @@ impl HexGridCell {
         &self,
         fill_color: ggez_gfx::Color,
         outline_color: ggez_gfx::Color,
-        radius: u8,
+        radius: usize,
         has_gradient: bool,
         resource_mesh_builder: &mut ggez_gfx::MeshBuilder
     ) {
